Fix veryisolated network without ip.
When using veryisolated network as a forwarding mode, the ip is not
required.  Currently there is no way to not provide an ip.  This should
fix it.

Furthermore the naming for the collision detection was producing names
like <net>0, <net>01, <net>012 and so on.  This switch to the more
efficient <net>0, <net>1, ... scheme.